Change return type for OpenApiSchemaVarType.number from 'float' to 'double'

+++ lib/src/config/open_api_generator_config.dart
@@ -47,7 +47,7 @@ class SwaggerToDartConfig {
           _ => 'String',
         };
       case OpenApiSchemaVarType.number:
-        return 'float';
+        return 'double';
       case OpenApiSchemaVarType.integer:
         return 'int';
       case OpenApiSchemaVarType.boolean:
